requirements.txt
streamlit
gemini.generativeai
PIL
python-dotenv